+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version alpha 0.0.3 (27/06/2023)
+
+- Removed debug middleware when not in debug mode
+
 ## Version alpha 0.0.2 (27/06/2023)
 
 - Added SPA routing without reloading